+++ src/main/java/uk/co/duelmonster/minersadvantage/common/feature/mining/ExcavationComponent.java
@@ -3,9 +3,8 @@ package uk.co.duelmonster.minersadvantage.common.feature.mining;
 import java.util.List;
 
 import uk.co.duelmonster.minersadvantage.common.component.ComponentLifecycle;
-    import uk.co.duelmonster.minersadvantage.common.component.ComponentTickHelper;
-    import uk.co.duelmonster.minersadvantage.common.config.ExcavationConfig;
-import uk.co.duelmonster.minersadvantage.common.registry.RegistryPredicates;
+import uk.co.duelmonster.minersadvantage.common.component.ComponentTickHelper;
+import uk.co.duelmonster.minersadvantage.common.config.ExcavationConfig;
 import uk.co.duelmonster.minersadvantage.common.services.mining.ExcavationCoreService;
 
 /**
@@ -98,14 +97,13 @@ public final class ExcavationComponent implements ComponentLifecycle {
 
         var context = ComponentTickHelper.getContext();
         if (service.isBlock(context.blockId())) {
-            int verticalRadius = RegistryPredicates.isShovelToolId(context.toolId()) ? 0 : config.radiusVertical();
             lastPlan = service.buildPlan(
                 context.blockX(),
                 context.blockY(),
                 context.blockZ(),
                 context.blockId(),
                 config.radiusHorizontal(),
-                verticalRadius,
+                config.radiusVertical(),
                 config.processesPerTick()
             );
         }